adicionando zod para validacao de inputs e confirmacao para caso queira deletar uma transacao

+++ src/components/TransactionItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Pressable } from "react-native";
+import { Alert, Pressable } from "react-native";
 import { Box } from "./ui/box";
 import { HStack } from "./ui/hstack";
 import { Text } from "./ui/text";
@@ -7,6 +7,7 @@ import { MaterialIcons } from "@expo/vector-icons";
 import { Transaction } from "../types";
 import { useTransactions } from "../hooks/useTransactions";
 import { EditTransactionModal } from "./EditTransactionModal"; // Importe o modal de edição
+import { View } from "react-native";
 
 interface TransactionItemProps {
   transaction: Transaction;
@@ -25,7 +26,25 @@ export const TransactionItem = ({
   const [isEditModalVisible, setIsEditModalVisible] = useState(false);
 
   const handleLongPress = () => {
-    deleteTransaction({ id: transaction.id });
+    // Exibe um alerta de confirmação antes de deletar
+    Alert.alert(
+      "Deletar Transação",
+      "Tem certeza que deseja deletar esta transação?",
+      [
+        {
+          text: "Cancelar",
+          style: "cancel", // Botão de cancelar
+        },
+        {
+          text: "Deletar",
+          style: "destructive", // Botão de deletar (estilo destrutivo)
+          onPress: () => {
+            // Se o usuário confirmar, deleta a transação
+            deleteTransaction({ id: transaction.id });
+          },
+        },
+      ]
+    );
   };
 
   const handleEditPress = () => {
@@ -52,11 +71,19 @@ export const TransactionItem = ({
                 size={20}
                 color={transaction.type === "incomes" ? "#4CAF50" : "#F44336"}
               />
-              <Text className="text-gray-800">{transaction.description}</Text>
-              <Pressable onPress={handleEditPress}>
-                <MaterialIcons name="edit" size={20} color="gray" />
-              </Pressable>
+              <View className="flex-1" style={{ maxWidth: "60%" }}>
+                <Text
+                  className="text-gray-800"
+                  numberOfLines={1} // Limita o texto a uma linha
+                  ellipsizeMode="tail" // Adiciona "..." ao final do texto
+                >
+                  {transaction.description}
+                </Text>
+              </View>
             </HStack>
+            <Pressable onPress={handleEditPress}>
+              <MaterialIcons name="edit" size={20} color="gray" />
+            </Pressable>
             <Text
               className={
                 transaction.type === "incomes"
